fix:  update expected folder structure (#22)
Each piece of data we receive in our `gcs_to_pubsub` function has a name
that gives us an idea of the folder structure of the bucket. Previously
we calculating the bucket name size, removing that from the data name,
and the figuring out the content type based on the positions of the
strings relative to the `/`.

It turns out the string we were splitting on never used the bucket name
to begin with so we were just removing a pre-determined number of
characters to validate the data we were getting. A change in the bucket
name increased the number of characters we were trimming which caused
most of our data to fail the length validation we do since it was
removing an extra `/`.

Since the `name` field doesn't use the bucket, I've changed it to just
split on the `/` and the indexing that follows all now works.